jira-agents/requirement.txt
langgraph
openai
fastapi
uvicorn
pydantic
httpx
python-dotenv
tiktoken